third_party/WebKit/LayoutTests/fast/css/invalidation/unresolved-pseudo.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
#sibling, #child, #child3 { background-color: green }
custom-one:unresolved #child,
custom-two:unresolved + #sibling,
custom-three:unresolved #child3 {
  background-color: red
}
</style>

<custom-one>
    <div></div>
    <div id="child"></div>
</custom-one>

<custom-two>
    <div></div>
    <div></div>
</custom-two>
<div id="sibling"></div>

<custom-three>
    <div></div>
    <div id="child3"></div>
</custom-three>

<script>
description("Use invalidation sets for :unresolved pseudo class.")

var red = "rgb(255, 0, 0)";
var green = "rgb(0, 128, 0)";

// Initially :unresolved.
shouldBe("getComputedStyle(child).backgroundColor", "red");
shouldBe("getComputedStyle(sibling).backgroundColor", "red");
shouldBe("getComputedStyle(child3).backgroundColor", "red");

document.body.offsetTop; // force recalc

var CustomOne = document.registerElement("custom-one", { prototype: Object.create(HTMLElement.prototype) });
document.head.appendChild(new CustomOne());

if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");
shouldBe("getComputedStyle(child).backgroundColor", "green");

document.body.offsetTop; // force recalc

var CustomTwo = document.registerElement("custom-two", { prototype: Object.create(HTMLElement.prototype) });
document.head.appendChild(new CustomTwo());

if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");
shouldBe("getComputedStyle(sibling).backgroundColor", "green");

document.body.offsetTop; // force recalc

customElements.define("custom-three", class extends HTMLElement {});
if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");
shouldBe("getComputedStyle(child3).backgroundColor", "green");

</script>
